requirements.txt
streamlit
python-dotenv
pathlib
google-generativeai
youtube_transcript_api